# lib/api-zod/src/generated/types/apiStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApiStatus {
  connected: boolean;
  /** @nullable */
  lastUpdated: Date | null;
  requestsInWindow: number;
  requestLimit: number;
  message: string;
}