# lib/api-zod/src/generated/types/elytraCategory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ElytraCategory = typeof ElytraCategory[keyof typeof ElytraCategory];


export const ElytraCategory = {
  elytra: 'elytra',
} as const;