# lib/api-zod/src/generated/types/elytraDashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApiStatus } from './apiStatus';
import type { ElytraStats } from './elytraStats';

export interface ElytraDashboard {
  stats: ElytraStats[];
  api: ApiStatus;
  qualifyingListings: number;
  generatedAt: Date;
}