# lib/api-zod/src/generated/types/elytraListing.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ElytraCategory } from './elytraCategory';

export interface ElytraListing {
  id: string;
  /** @nullable */
  itemId: string | null;
  displayName: string;
  category: ElytraCategory;
  enchantments: string[];
  price: number;
  seller: string;
  /** @nullable */
  sellerUuid: string | null;
  quantity: number;
  /** @nullable */
  timeRemaining: string | null;
  collectedAt: Date;
}