# lib/api-zod/src/generated/types/elytraStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ElytraCategory } from './elytraCategory';

export interface ElytraStats {
  category: ElytraCategory;
  /** @nullable */
  lowest: number | null;
  /** @nullable */
  highest: number | null;
  /** @nullable */
  average: number | null;
  /** @nullable */
  median: number | null;
  activeListings: number;
  /** @nullable */
  priceChange: number | null;
  currency: string;
}