# lib/api-zod/src/generated/types/elytraTransaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ElytraCategory } from './elytraCategory';

export interface ElytraTransaction {
  id: string;
  seller: string;
  price: number;
  category: ElytraCategory;
  enchantments: string[];
  quantity: number;
  timestamp: Date;
}