# lib/api-zod/src/generated/types/marketAlert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ElytraCategory } from './elytraCategory';
import type { MarketAlertType } from './marketAlertType';

export interface MarketAlert {
  id: number;
  type: MarketAlertType;
  category: ElytraCategory;
  affectedQuantity: number;
  /** @nullable */
  previousPrice: number | null;
  /** @nullable */
  currentPrice: number | null;
  /** @nullable */
  percentageChange: number | null;
  /** @nullable */
  estimatedValue: number | null;
  detectedAt: Date;
}