# lib/api-zod/src/generated/types/priceObservation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ElytraCategory } from './elytraCategory';

export interface PriceObservation {
  id: number;
  category: ElytraCategory;
  timestamp: Date;
  price: number;
  open: number;
  high: number;
  low: number;
  close: number;
  /** @nullable */
  priceChange: number | null;
  sampleSize: number;
  observationCount: number;
}