# lib/api-zod/src/generated/types/timeRange.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TimeRange = typeof TimeRange[keyof typeof TimeRange];


export const TimeRange = {
  five_minutes: 'five_minutes',
  hour: 'hour',
  today: 'today',
  seven_days: 'seven_days',
  thirty_days: 'thirty_days',
  one_year: 'one_year',
} as const;